casting cores to int

+++ pipeline/pipeline/pipeline.py
@@ -176,7 +176,7 @@ class submitWorker(object):
         kwargs = dict(test=self.test)
         if self.weights is not None:
             try: 
-                kwargs['cores'] = self.weights[filt] * self.stage.pipeline['total_cores']
+                kwargs['cores'] = int(self.weights[filt] * self.stage.pipeline['total_cores'])
             except KeyError:
                 pass
